refactor existing frontend
Rearrange existing parts of frontend
