+++ FrontEnd/src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
-import { AuthRoutingModule } from './auth/auth-routing.module';
-import { ErrorsRoutingModule } from './auth/errors-routing.module';
+import { AuthRoutingModule } from './core/auth/auth-routing.module';
+import { ErrorsRoutingModule } from './core/auth/errors-routing.module';
 import { HomeRoutingModule } from './home/home-routing.module';
 
 @NgModule({
