+++ FrontEnd/src/app/core/auth/components/register/register.component.ts
