+++ FrontEnd/src/app/core/auth/enums/roles.enum.ts
