+++ FrontEnd/src/app/core/auth/errors/error-404/error-404.component.ts
