+++ FrontEnd/src/app/core/auth/forms/change-password.form.ts
@@ -1,6 +1,6 @@
 import { Injectable } from "@angular/core";
 import { Validators } from "@angular/forms";
-import { BaseFormModel } from "src/custom-elements/controls/base-form-model/base-form-model";
+import { BaseFormModel } from "src/app/shared/models/base-form.model";
 import { ChangePasswordModel } from "../models/change-password.model";
 
 @Injectable()
