--- FrontEnd/src/app/navbar/navbar.module.ts
@@ -1,21 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
-import { AuthModule } from '../auth/auth.module';
-import { NavbarComponent } from './navbar.component';
-
-@NgModule({
-  declarations: [
-    NavbarComponent
-  ],
-  imports: [
-    CommonModule,
-    RouterModule,
-    AuthModule
-  ],
-  exports: [
-    NavbarComponent
-  ],
-  providers: [],
-})
-export class NavbarModule { }
