+++ FrontEnd/src/app/shared/shared.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from "@angular/core";
+
+
+@NgModule({
+  imports: [
+  ],
+  declarations: [
+  ],
+  exports: [
+  ],
+})
+export class SharedModule {}
